OAuth2: remove access token cookie from successful authentication response
No need to send the access token back to the front end. We'll set a session cookie that consumers
of this application use only to access this back channel; the back end will use the access token
on behalf of the user, without having to expose it to a less secure front channel.